requirements.txt
streamlit
boto3
pandas
requests
Pillow